Update PHPUnit to latest version, move current tests into `old` directory
