+++ composer.json
@@ -25,7 +25,7 @@
     "symfony/serializer": "^3.1"
   },
   "require-dev": {
-    "phpunit/phpunit": "4.6.*",
+    "phpunit/phpunit": "^7.3",
     "keboola/coding-standard": "^2.0",
     "symfony/finder": "*",
     "phpstan/phpstan": "^0.8"
@@ -41,7 +41,10 @@
     }
   },
   "scripts": {
-    "tests": "phpunit",
+    "old-tests": "phpunit tests/Old",
+    "tests": [
+      "@old-tests"
+    ],
     "phpstan": "phpstan analyse --no-progress --level=max ./src ./tests -c phpstan.neon",
     "phpcs": "phpcs -n --ignore=vendor --extensions=php .",
     "phpcbf": "phpcbf -n --ignore=vendor --extensions=php .",
